+++ @xen-orchestra/web/src/components/alarms/AlarmLink.vue
