+++ @xen-orchestra/web/src/pages/host/[id]/dashboard.vue
@@ -14,7 +14,12 @@
       </VtsStateHero>
     </div>
     <template v-else>
-      <HostDashboardAlarms class="alarms" :host />
+      <DashboardAlarms
+        class="alarms"
+        :alarms="hostAlarms"
+        :is-ready="areHostAlarmsReady"
+        :has-error="hasHostAlarmFetchError"
+      />
       <HostDashboardPatches class="patches" :host />
       <HostDashboardVmsStatus class="vms-status" :host />
       <HostDashboardCpuProvisioning class="cpu-provisioning" :host />
@@ -28,7 +33,7 @@
 </template>
 
 <script lang="ts" setup>
-import HostDashboardAlarms from '@/components/host/dashboard/alarms/HostDashboardAlarms.vue'
+import DashboardAlarms from '@/components/alarms/DashboardAlarms.vue'
 import HostDashboardCpuProvisioning from '@/components/host/dashboard/HostDashboardCpuProvisioning.vue'
 import HostDashboardCpuUsageChart from '@/components/host/dashboard/HostDashboardCpuUsageChart.vue'
 import HostDashboardLoadAverageChart from '@/components/host/dashboard/HostDashboardLoadAverageChart.vue'
@@ -39,6 +44,7 @@ import HostDashboardRamProvisioning from '@/components/host/dashboard/HostDashbo
 import HostDashboardRamUsageChart from '@/components/host/dashboard/HostDashboardRamUsageChart.vue'
 import HostDashboardVmsStatus from '@/components/host/dashboard/HostDashboardVmsStatus.vue'
 import { useFetchStats } from '@/composables/fetch-stats.composable.ts'
+import { useXoHostAlarmsCollection } from '@/remote-resources/use-xo-host-alarms-collection.ts'
 import { type XoHost } from '@/types/xo/host.type'
 import { GRANULARITY } from '@/utils/rest-api-stats.ts'
 import VtsStateHero from '@core/components/state-hero/VtsStateHero.vue'
@@ -52,6 +58,8 @@ const { host } = defineProps<{
 const { t } = useI18n()
 
 const { data, isFetching, error } = useFetchStats('host', () => host.id, GRANULARITY.Hours)
+
+const { hostAlarms, areHostAlarmsReady, hasHostAlarmFetchError } = useXoHostAlarmsCollection({}, () => host.id)
 
 const uiStore = useUiStore()
 </script>
@@ -91,6 +99,7 @@ const uiStore = useUiStore()
 
   .alarms {
     grid-area: alarms;
+    height: 46.2rem;
   }
 
   .patches {
